side menu has been split into two
